requirements.txt
azure-functions==1.18.0
fastmcp==0.1.0
aiohttp==3.9.0
pandas==2.1.0
azure-storage-blob==12.19.0
pydantic==2.5.0
applicationinsights==0.11.10
pytest==7.4.3
pytest-asyncio==0.21.1